requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
